+++ nexus-webapp/pom.xml
@@ -327,12 +327,7 @@
               <includes>
                 <!-- some order required here -->
                 <include>lib/string.js</include>
-                <include>lib/nx.js</include>
-                <include>lib/nx.assert.js</include>
-                <include>lib/nx.log.js</include>
-                <include>lib/nx.define.js</include>
                 <include>lib/require-*.js</include>
-                <include>lib/classes/**/*.js</include>
               </includes>
             </configuration>
           </execution>
