+++ nexus-webapp/src/main/webapp/js/NX/LegacyLogMixin.js
@@ -10,43 +10,20 @@
  * of Sonatype, Inc. Apache Maven is a trademark of the Apache Software Foundation. M2eclipse is a trademark of the
  * Eclipse Foundation. All other trademarks are the property of their respective owners.
  */
-
-/*global Ext,NX */
-
-Ext.ns('NX');
-
-/**
- * Exception thrown when assertion is triggered.
- *
- * @param message
- * @constructor
- *
- * @since 2.4
- */
-function AssertError(message) {
-    this.name = 'AssertError';
-    this.message = message || 'Assertion failure';
-}
-AssertError.prototype = new Error();
-AssertError.prototype.constructor = AssertError;
-
-/**
- * Assertion helpers.
- */
-Ext.apply(NX, {
-
-    /**
-     * Assert a condition.  When the condition expression is false and exception is thrown.
-     *
-     * @param expression    Condition expression
-     * @param message       Exception message
-     *
-     * @since 2.4
-     */
-    assert: function (expression, message) {
-        if (!expression) {
-            throw new AssertError(message);
-        }
-    }
-
+/*global define,NX*/
+define('NX/LegacyLogMixin', ['NX/log','NX/define'], function() {
+  /**
+   * Legacy $log() mixin.
+   *
+   * @since 2.4
+   */
+  NX.define('NX.LegacyLogMixin', {
+      /**
+       * @deprecated
+       */
+      '$log': function(message) {
+          NX.log.debug(this.$className + ': ' + message);
+          NX.log.warn('[DEPRECATED]', this.$className + '.$log() is deprecated; use NX.LogAwareMixin instead');
+      }
+  });
 });
