+++ nexus-webapp/src/main/webapp/js/NX/bootstrap.js
@@ -10,20 +10,11 @@
  * of Sonatype, Inc. Apache Maven is a trademark of the Apache Software Foundation. M2eclipse is a trademark of the
  * Eclipse Foundation. All other trademarks are the property of their respective owners.
  */
-
-/**
- * Legacy $log() mixin.
- *
- * @since 2.4
- */
-NX.define('NX.LegacyLogMixin', {
-
-    /**
-     * @deprecated
-     */
-    '$log': function(message) {
-        NX.log.debug(this.$className + ': ' + message);
-        NX.log.warn('[DEPRECATED]', this.$className + '.$log() is deprecated; use NX.LogAwareMixin instead');
-    }
-
-});
+/*global define*/
+define('NX/bootstrap', [
+  'NX/base',
+  'NX/define',
+  'NX/log',
+  'NX/LegacyLogMixin',
+  'NX/LogAwareMixin'
+]);
